+++ src/Validator/Allowed.php
@@ -6,6 +6,7 @@ namespace Duon\Sire\Validator;
 
 use Duon\Sire\Contract;
 use Duon\Sire\DslSplitter;
+use Duon\Sire\Validation;
 use Override;
 
 /** @api */
@@ -14,10 +15,10 @@ final class Allowed implements Contract\Validator
 	public string $message = 'Invalid value';
 
 	#[Override]
-	public function validate(Contract\Value $value, string ...$args): bool
+	public function validate(Contract\Value $value, string ...$args): Contract\Validation
 	{
 		$allowed = DslSplitter::split($args[0] ?? '', ',');
 
-		return in_array($value->value, $allowed, true);
+		return Validation::from(in_array($value->value, $allowed, true));
 	}
 }