+++ src/Validator/MaxLength.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Duon\Sire\Validator;
 
 use Duon\Sire\Contract;
+use Duon\Sire\Validation;
 use Override;
 
 /** @api */
@@ -13,8 +14,8 @@ final class MaxLength implements Contract\Validator
 	public string $message = 'Exeeds the maximum length of %4$s characters';
 
 	#[Override]
-	public function validate(Contract\Value $value, string ...$args): bool
+	public function validate(Contract\Value $value, string ...$args): Contract\Validation
 	{
-		return strlen($value->value) <= (int) ($args[0] ?? null);
+		return Validation::from(strlen($value->value) <= (int) ($args[0] ?? null));
 	}
 }