+++ src/Validator/Required.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Duon\Sire\Validator;
 
 use Duon\Sire\Contract;
+use Duon\Sire\Validation;
 use Override;
 
 /** @api */
@@ -13,18 +14,18 @@ final class Required implements Contract\ValidatesEmpty
 	public string $message = 'Required';
 
 	#[Override]
-	public function validate(Contract\Value $value, string ...$args): bool
+	public function validate(Contract\Value $value, string ...$args): Contract\Validation
 	{
 		$val = $value->value;
 
 		if (is_null($val)) {
-			return false;
+			return Validation::invalid();
 		}
 
 		if (is_array($val) && count($val) === 0) {
-			return false;
+			return Validation::invalid();
 		}
 
-		return true;
+		return Validation::valid();
 	}
 }